Safari-611.1.10/LayoutTests/accessibility/mac/select-text/select-text-135546-expected.txt

The quick brown fox jumps over the lazy dog.

TEXT2: The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.

This tests the ability to select and replace text with respect to selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 'quick'
PASS windowSelection.getRangeAt(0).startOffset is 11
PASS windowSelection.getRangeAt(0).endOffset is 16
PASS result is 'quick'
PASS windowSelection.getRangeAt(0).startOffset is 56
PASS windowSelection.getRangeAt(0).endOffset is 61
PASS successfullyParsed is true

TEST COMPLETE
